Added Toggle
I added a toggle to make players more in control of their rp, sometimes they don't always want a weapon on their back. Pressing left alt with a valid weapon in inventory will place the weapon on their back, pressing left alt again will remove the weapon. I also took the liberty of organizing the code a bit to make it intuitive to people who want to mess with the code.